refactor(gateway): extract LlmRouterConfig (GW-2 S4)
